[feat] 좋아요 순으로 리스트 정렬하도록 수정

+++ kitten-api/src/main/java/kitten/diy/api/application/port/in/query/data/BoardInfoData.java
@@ -30,7 +30,7 @@ public record BoardInfoData(
                 .boardKey(data.boardKey())
                 .imageUrl(data.imageUrl())
                 .createTime(data.createTime())
-                .likeCount(data.likeCount())
+                .likeCount(data.likeCount().intValue())
                 .viewCount(data.viewCount())
                 .build();
     }
